requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7  # Para documentación Swagger/OpenAPI
drf-spectacular==0.26.5  # Alternativa moderna a drf-yasg
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0
gunicorn==21.2.0
django-model-utils==4.3.1
python-slugify==8.0.1
Pillow==10.0.0
celery==5.3.1
redis==4.5.5
django-extensions==3.2.3
django-filter==23.3
# Usando la versión estable de django-ckeditor
django-ckeditor==6.7.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
